feat: Setting up the project bones (#1)
* feat(ui-header): Create component

Define the base structure of the header component as part of the “project bones” step.

* fix(ui-content): Create component

Define the base structure of the content component as part of the “project bones” step. This includes:
- Required sidebars
- Main content area
- Ability to show/hide the secondary sidebar for elements

* fix: Clean-up main styles

* fix(app): Call new components

Update the app main content with its bone structure by adding:
- Header component
- Content component

* feat: Upload placeholder image

* fix: Remove unused components

* fix(ui-content): Define base elements

* feat(ui-content): Add drag and drop ability

* fix(ui-content): Define image element width

* fix(ui-content): Improve event handling

* fix(ui-content): Improve component reactivity

Encourage the use of Composition API for state management by refactoring `data` and `methods` using `ref`.

* fix(ui-component): Simplify dynamic classes

* feat(ui-component): Add export content functionality

- Convert the `dropppedElements` array to JSON format.
- Show in the console the JSON content.

* feat(ui-component): Download fields as JSON file

This improvement allows the export button to trigger the download of a JSON file (`content.json`) with the elements available in the dropzone. It is the same content appearing in the console.

* feat(ui-content): Reorganize elements

Enable reorder functionality for the `droppedElements`.

* feat(ui-content): Remove elements functionality

- Add a “remove” button for each dropped element.
- Create a method to remove the element from the `droppedElements` array.

* feat: Install `sass-embedded`

Install `sass-embedded` to support embedded SCSS

* feat(ui-sidebar): Create new component

- Create a simple component for the sidebar that supports title and child content.
- Update `UIContent` using the new `UISidebar` component.

* feat(ui-button): Create new component

Create a simple button component.

* feat(ui-sidebar): Add floating variation

Define `float` property to allow a floating version of the component.

* fix(ui-content): Update template

- Update content to use the `UISidebar` and `UIButton` components.
- Move secondary sidebar inside the main left sidebar to create a floating submenu and avoid pushing main content every time it appears.

* fix(ui-sidebar): Improve floating styles

* feat(ui-menu): Create new component

* fix(ui-content): Use `UIMenu` component

* feat(block-wrapper): Create new component

This component intention is to provide common wrapper for all the items added to the drop area.

* feat(block-text): Create new component

This component was created to represent the text element.

* feat(block-image): Create new component

This component was created to represent the image element.

* feat(block-main): Create new component

This is the main component for blocks. Depending on the “element” it displays the correct block element content (markup).

* fix(ui-sidebar): Give styles to the inner list

* Install `vuedraggable` package

* fix(ui-content): Refactor drag and drop experience

In order to handle smoother events, I replaced the basic `draggable` actions with `draggableComponent` for Vue 3. This commit also includes:
- Update in the markup
- Update in the styles
- Move from Vue-specific options to a defining component
- Reorder elements using up or down button
- Use new components created for better organization
- Use SCSS lang for styles

* fix(ui-content): General improvements

- Remove unnecessary class.
- Close sidebar when element drops in the drop zone.
- Enable image element options.
- Add reference styles for ghost element.
- Resize canvas width based on 3 breakpoints: desktop, tablet, mobile.

* feat(builder-mask): Create new component

Originally the project used a builder mask. In the meantime, it is not necessary, but just in case we need it in the future it was converted into a single component.

REMOVE when cleaning up the code and ensuring this component is not being used anywhere in the project.

* feat(block-text): Implement inline text editor

This commit contains different changes to enable an inline editor for the `BlockText` component.

- Install `quill` editor.
- Install `vue-quilly` component.
- Wrap `BlockText` value with Quilly Editor.
- Remove Quilly Editor toolbar. This decision will simplify the content edition.
- Define an `id` property for the block components. This will help Quilly link the editor changes to an specific component.
- Ensure the value of the block updates when changes happen.

* fix(block-text): Ensure text updated passes to array

This commit resolves a problem with the `BlockText` value propagation.

- Use `@update:modelValue` property from `QuillyEditor` instead of `watch` to update content.
- Setting up `handleUpdate` prop on `BlockMain`.
- Ensure updated block uses splice to pick up changes on `UIContent`.