feat: Lazy load on start (#9)
* fix: Improve core styles with smaller tokens

Use TailwindCSS idea to configure necessary tokens to style faster the app.

Why not TailwindCSS?
TailwindCSS is a great option; however, since the project is small it won’t be necessary.

* feat(ui-loading): Create new component

This commit creates a loading component that will run on top of the app when it starts; while the rest of the assets and blocks load.