fix: Code clean-up and improvement (#10)
* fix(app): Clean-up and simplify code

* fix: Move `asyncLoadComponent` to utils folder

* fix(builder-elements): Clean-up and simplify code

- Move elements to its own reusable component.
- Include the `@element` alias.

* fix(builder-header): Clean-up and simplify code

This commit granulates the code making it mode reusable and maintainable for the future by creating two new components:
- `HeaderBar`
- `HeaderMenu`

It also register a new alias `@header` to find elements in the folder faster.

* fix(builder-sidebar): Clean-up and simplify code

This commit remove commented-out code and adds the `scoped` attribute to the `style` tag.

* fix(builder-main): Clean-up and simplify methods

This commit improves the methods avoiding code repetition.