fix: Improve “add element” through click on mobiles
This commit fixes a reactivity problem between `dropped-elements` component that’s inside the `BuilderMain` component and the elements added through click event from the mobiles menu bar.

To resolve the issue the new element added is passed to the `BuilderMain` through `addedElement` prop. This property gets watched and updates accordingly.