fix: Migrate to MSW for API Mocking (#12)
* Install and configure MSW

* fix(app): Fetch default elements from API

* fix(app): Fetch dropped elements

* fix: Mock handler for `http.put` to receive payload

* fix: Create request handlers to put, patch, and delete on `canvas` API

* fix(builder-main): Integrate with API canvas

* fix(builder-main): Removing incorrect element

* fix(builder-main): Fetch real data from API/canvas

At the previous implementation, I was simulating the fetch of the data for testing purposes.

In this commit, I start fetching real data and log changes on the console to debug if data is being updated correctly.

* fix(app): PUT new element on canvas

This commit ensures an element added through click button (on mobiles) successfully updates data on API canvas.

* fix(app): Export data from API canvas

This commit ensures the data exported and logged to the console come from the API canvas.

* fix(handlers): Add element to API canvas with unique ID

* fix: Stop fetching API canvas on mount

This commit removes onMount fetching for dropped elements to the API canvas because of two important reasons:
1. The fetch was inserting an empty item to the array unnecessary.
2. There is no need to fetch API canvas onMount since API canvas is empty. There are no predefined elements added to it on load.

* fix(handlers): Update PUT method for API canvas

* test(builder-main): Temporarilty remove try-catch syntax

This commit removes temporarily the try-catch syntax from the code to focus on the API requests rather than error handling.

In a later commit the try-catch syntax can return to provide a more structured way to handle erros.

* fix(builder-main): Stop updating API canvas on every interaction

* feat(ui-devices): Create new component

This commit create a new component for the devices element allowing to have a cleaner code and more mantainable.

It also includes the `@admin` alias. This alias will replace in the future `@Block`, `@Field`, and all the other unnecessary alias cleaning up the alias structure a bit more.

* fix(handler): Replace instead of pushing

This commit fixes an issue with the PUT handler to ensure it replaces `canvasData` instead of pushing to it.

* feat: Install axios package

* fix(builder-main): Push to API canvas on element add

This commit moves the original `BuilderMain` code to a new file: `BuilderMainOld.vue` in order to resolve some issues when syncing draggable components and API canvas.

The `BuilderMain` file includes the following improvements:
- Remove of the old code and start fresh.
- Include base structure and draggable-component.
- Include `UIDevices` component.
- Fetch and sync API canvas.
- Include a method that will trigger when user adds a new element to the drop zone. This method will update the droppedElements and the API canvas.

* fix(builder-main): Remove duplicated variable

This commit removes the `droppedElements` variable since the file already has the `elements` variable declared making redundant and unnecessary keeping both.

* fix: Improve “add element” through click on mobiles

This commit fixes a reactivity problem between `dropped-elements` component that’s inside the `BuilderMain` component and the elements added through click event from the mobiles menu bar.

To resolve the issue the new element added is passed to the `BuilderMain` through `addedElement` prop. This property gets watched and updates accordingly.

* fix(builder-main): Set elements with updated value

This commit brings back the `BlockMain` component and ensures the component value gets updated in the API canvas.

* fix(builder-main): Set container height for `BlockMain`

* feat(settings-container): Create new component

This commit splits the element’s “container settings” in its own file for better maintenance in the future and allows user to toggle each option visibility based on the element needs.

* feat(settings-image): Create new component

This commit splits the element’s “image settings” to its own file for better maintenance in the future.

* fix(builder-main): Update canvas on sort

This commit brings back the `BlockTools` component and improves the sorting functionality by updating the API canvas.

It also renames the `addElement` method to `updateCanvas`.

* fix(builder-main): Update canvas on copy

This commit brings back the “copy” functionality for the `BlockTools` component. It also updates the API canvas to ensure proper syncing.

* fix(builder-main): Update canvas on delete

This commit brings back the “remove” functionality for the `BlockTools` component. It also updates the API canvas to ensure proper syncing.

* fix(builder-main): Update canvas on reorganize

This commit brings back the “move-up” and “move-down” functionality for the `BlockTools` component. It also updates the API canvas to ensure proper syncing.

* fix(builder-main): Toggle element tools on focus and blur

* fix(builder-main): Enable edit on toolbar

This commit focuses on the toggle functionality for the elements edit tool. It only opens or closes, depending the case, and passes the title for the settings box.

* fix(builder-main): Enable element settings for container

* fix(builder-main): Move back “image” settings

There is a problem with reactivity with `SettingsImage` component when it contains the `UIRadioImageGroup`.

As a momentary fix and while investigate, the image settings will be back to the `BuilderMain` component.

* fix(builder-main): Update element container props in API canvas

* fix(builer-main): Incorrect update of element properties

* fix(builder-main): Update image on API canvas

* fix: Remove old file

This commit removes the old copy of `BuilderMain` component.

* fix(builder-main): Incorrect update of methods

This commit fixes an incorrect update of the methods. By mistake, the `updateElement` method was updated instead of the `updateContainer`.

* styles: Fix linter issues